Move chainId type to long and add palm

+++ app/src/main/java/com/alphawallet/app/entity/NetworkInfo.java
@@ -20,7 +20,7 @@ public class NetworkInfo extends com.alphawallet.ethereum.NetworkInfo {
             String symbol,
             String rpcServerUrl,
             String etherscanUrl,
-            int chainId,
+            long chainId,
             String backupNodeUrl,
             String etherscanAPI) {
         super(name, symbol, rpcServerUrl, etherscanUrl, chainId);
