+++ app/src/main/java/com/alphawallet/app/router/SendTokenRouter.java
@@ -11,7 +11,7 @@ import com.alphawallet.app.entity.tokens.Token;
 import com.alphawallet.app.ui.SendActivity;
 
 public class SendTokenRouter {
-    public void open(Activity context, String address, String symbol, int decimals, Wallet wallet, Token token, int chainId) {
+    public void open(Activity context, String address, String symbol, int decimals, Wallet wallet, Token token, long chainId) {
         Intent intent = new Intent(context, SendActivity.class);
         intent.putExtra(C.EXTRA_CONTRACT_ADDRESS, address);
         intent.putExtra(C.EXTRA_ADDRESS, token.getAddress());
