+++ app/src/main/java/com/alphawallet/app/ui/RedeemAssetSelectActivity.java
@@ -71,7 +71,7 @@ public class RedeemAssetSelectActivity extends BaseActivity implements OnTokenCl
         viewModel = new ViewModelProvider(this, viewModelFactory)
                 .get(RedeemAssetSelectViewModel.class);
 
-        int chainId = getIntent().getIntExtra(C.EXTRA_CHAIN_ID, EthereumNetworkBase.MAINNET_ID);
+        long chainId = getIntent().getLongExtra(C.EXTRA_CHAIN_ID, EthereumNetworkBase.MAINNET_ID);
         token = viewModel.getTokensService().getToken(chainId, getIntent().getStringExtra(C.EXTRA_ADDRESS));
         ticketRange = getIntent().getParcelableExtra(TICKET_RANGE);
         setContentView(R.layout.activity_redeem_asset);
